Move exiled governments decisions to their own file
