+++ Vic2ToHoI4/Source/HOI4World/Decisions/Decisions.h
@@ -6,6 +6,7 @@
 #include "IdeologicalDecisions.h"
 #include "DecisionsCategories.h"
 #include "DecisionsCategory.h"
+#include "ExiledGovernmentsDecisions.h"
 #include "PoliticalDecisions.h"
 #include "StabilityWarSupportDecisions.h"
 #include "../../Configuration.h"
@@ -45,7 +46,7 @@ class decisions: commonItems::parser
 		}
 		[[nodiscard]] const std::vector<decisionsCategory>& getExiledGovernmentsDecisions() const
 		{
-			return exiledGovernmentsDecisions;
+			return exiledGovernmentsDecisions.getDecisions();
 		}
 		[[nodiscard]] const std::vector<decisionsCategory>& getForeignInfluenceDecisions() const
 		{
@@ -58,7 +59,6 @@ class decisions: commonItems::parser
 		[[nodiscard]] const std::vector<decisionsCategory>& getGenericDecisions() const { return genericDecisions; }
 
 	private:
-		void updateExiledGovernmentDecisions(const std::set<std::string>& majorIdeologies);
 		void updateForeignInfluenceDecisions(const std::set<std::string>& majorIdeologies);
 		void updateMtgNavalTreatyDecisions(const std::set<std::string>& majorIdeologies);
 		void updateGenericDecisions(
@@ -70,7 +70,7 @@ class decisions: commonItems::parser
 
 		StabilityWarSupportDecisions stabilityDecisions;
 		PoliticalDecisions politicalDecisions;
-		std::vector<decisionsCategory> exiledGovernmentsDecisions;
+		ExiledGovernmentsDecisions exiledGovernmentsDecisions;
 		std::vector<decisionsCategory> foreignInfluenceDecisions;
 		std::vector<decisionsCategory> mtgNavalTreatyDecisions;
 		std::vector<decisionsCategory> genericDecisions;
