+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj
@@ -214,6 +214,7 @@
     <ClCompile Include="Source\HOI4World\Decisions\Decisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\DecisionsCategory.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\DecisionsFile.cpp" />
+    <ClCompile Include="Source\HOI4World\Decisions\ExiledGovernmentsDecisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\IdeologicalDecisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\PoliticalDecisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\StabilityWarSupportDecisions.cpp" />
@@ -352,6 +353,7 @@
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsCategories.h" />
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsCategory.h" />
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsFile.h" />
+    <ClInclude Include="Source\HOI4World\Decisions\ExiledGovernmentsDecisions.h" />
     <ClInclude Include="Source\HOI4World\Decisions\IdeologicalDecisions.h" />
     <ClInclude Include="Source\HOI4World\Decisions\PoliticalDecisions.h" />
     <ClInclude Include="Source\HOI4World\Decisions\StabilityWarSupportDecisions.h" />
