+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj.filters
@@ -459,6 +459,9 @@
     <ClCompile Include="Source\HOI4World\Decisions\PoliticalDecisions.cpp">
       <Filter>HoI4World\Decisions</Filter>
     </ClCompile>
+    <ClCompile Include="Source\HOI4World\Decisions\ExiledGovernmentsDecisions.cpp">
+      <Filter>HoI4World\Decisions</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\common_items\Date.h">
@@ -873,6 +876,9 @@
       <Filter>HoI4World\Decisions</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsFile.h">
+      <Filter>HoI4World\Decisions</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Decisions\ExiledGovernmentsDecisions.h">
       <Filter>HoI4World\Decisions</Filter>
     </ClInclude>
   </ItemGroup>
